+++ lib/analysis/ai/memory.rs
@@ -8,12 +8,13 @@ use error::*;
 use memory::paged;
 use memory;
 use RC;
+use serde::Serialize;
 use types::Endian;
 
 
 /// The `Value` trait must be implemented for abstract values operated over by
 /// this memory model.
-pub trait Value: memory::value::Value + domain::Value {}
+pub trait Value: memory::value::Value + domain::Value + Serialize {}
 
 
 /// A memory model for abstract interpretation.
@@ -105,6 +106,13 @@ impl<'m, V> Memory<'m, V> where V: Value {
             self.0.pages.insert(*other_page.0, page);
         }
         Ok(self)
+    }
+}
+
+
+impl<'m, V: domain::Value + Value> domain::Memory<V> for Memory<'m, V> {
+    fn join(self, other: &Memory<V>) -> Result<Memory<'m, V>> {
+        self.join(other)
     }
 }